Solution to Check Leap Year

+++ Check Leap Year@@6788dd63490637cfc44e1b90/code.c
@@ -1,8 +1,8 @@
 // Your code here...
 #include<stdio.h>
 int main(){
-    double x;
-    scanf("%lf",x);
+    int x;
+    scanf("%d",x);
     if((x%4==0 && x%4==100) || x%400==0){
         printf("Leap Year");
     }
